--- web/jest.config.js
@@ -1,17 +1,0 @@
-const nextJest = require('next/jest')
-
-const createJestConfig = nextJest({ dir: './' })
-
-const config = {
-  testEnvironment: 'jsdom',
-  setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
-  moduleNameMapper: {
-    '^@/(.*)$': '<rootDir>/src/$1',
-  },
-  testMatch: [
-    '**/test/**/*.[jt]s?(x)',
-    '**/?(*.)+(spec|test).[jt]s?(x)',
-  ],
-}
-
-module.exports = createJestConfig(config)
