+++ web/jest.config.ts
@@ -1,0 +1,15 @@
+import type { Config } from 'jest'
+import nextJest from 'next/jest'
+
+const createJestConfig = nextJest({ dir: './' })
+
+const config: Config = {
+  testEnvironment: 'jsdom',
+  setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
+  moduleNameMapper: {
+    '^@/(.*)$': '<rootDir>/src/$1',
+  },
+  testMatch: ['**/test/**/*.[jt]s?(x)', '**/?(*.)+(spec|test).[jt]s?(x)'],
+}
+
+export default createJestConfig(config)
