+++ web/package.json
@@ -66,6 +66,7 @@
     "prettier": "3.8.3",
     "tailwindcss": "4",
     "ts-jest": "29.4.10",
+    "ts-node": "^10.9.2",
     "typescript": "5"
   },
   "packageManager": "pnpm@9.15.9"
