+++ web/tsconfig.json
@@ -28,7 +28,8 @@
     "**/*.tsx",
     ".next/types/**/*.ts",
     ".next/dev/types/**/*.ts",
-    "**/*.mts"
-, "jest.config.js"  ],
+    "**/*.mts",
+    "jest.config.ts"
+  ],
   "exclude": ["node_modules"]
 }
